Production server

+++ backend/files/start.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+pip install -r requirements.txt
+
 wait_for_db() {
   echo "Waiting for database connection..."
   while ! nc -z db 5432; do
@@ -16,32 +18,16 @@ wait_for_db
 python manage.py makemigrations
 python manage.py migrate
 
-# Delete all existing superusers
-# python manage.py shell -c "from django.contrib.auth.models import User; User.objects.filter(is_superuser=True).delete()"
-python manage.py shell -c "from django.contrib.auth import get_user_model; User = get_user_model(); User.objects.filter(is_superuser=True).delete()"
-
-# Create the superuser from .env file
-python manage.py createsuperuser --noinput
-
-# #create test user
-# python manage.py createsuperuser --noinput --username=johndoe
-# python manage.py createsuperuser --noinput --username=johndoe2
-# python manage.py createsuperuser --noinput --username=johndoe3
-# python manage.py createsuperuser --noinput --username=pnolte
-# python manage.py createsuperuser --noinput --username=albert
-# python manage.py createsuperuser --noinput --username=kathrin
-# python manage.py createsuperuser --noinput --username=kekse
-# python manage.py createsuperuser --noinput --username=kek
-# python manage.py createsuperuser --noinput --username=nikkka
-# python manage.py createsuperuser --noinput --username=stephanie
+# check if DJANGO_SETTINGS_MODULE is set to production or development
+if [ "$DJANGO_SETTINGS_MODULE" = "backend.settings.development" ]; then
+  # Delete all existing superusers and create a new one from .env file
+  python manage.py shell -c "from django.contrib.auth import get_user_model; User = get_user_model(); User.objects.filter(is_superuser=True).delete()"
+  python manage.py createsuperuser --noinput
+fi
 
 # Collect static files (needed for serving static files with daphne)
 python manage.py collectstatic --noinput
 
 # Start the ASGI server (Daphne in this case)
 daphne -b 0.0.0.0 -p 8000 backend.asgi:application
-
 # tail -f /dev/null
-
-# Create an app
-# python manage.py startapp custom_auth
